+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './newClient';
+export * as mvm from './mvm';
