+++ src/newClient/ws.ts
@@ -1,6 +1,6 @@
 import WebSocket from 'ws';
 import Keystore from './types/keystore';
-import { signAuthenticationToken } from './utils/auth';
+import { signAuthenticationToken } from './utils/signToken';
 
 export function websocket(keystore: Keystore, url: string): WebSocket {
   const jwtToken = signAuthenticationToken('GET', '/', '', keystore) || '';
